// src/openapi/core.d.ts
export interface paths {
// ...
    "/core/access/v1/namespaces": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Gets the list of namespaces the current user can access. If the current user is able to access the
         *     'All Namespaces' option that will be indicated in the response. */
        get: operations["accessGetNamespaces"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/core/alarm/v2/alarms": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Gets all alarms including namespaced alarms from all namespaces and all non-namespaced
         *     alarms.  This API can get a snapshot of the alarms, or can return a
         *     stream of the alarms and the changes to them. The API does not provide
         *     suppressed alarms by default, but the suppressed alarms will be included if
         *     the "all" query parameter is set to "true". */
        get: operations["alarmGetAlarms"];
        /** @description Performs the specified acknowledgement action provided in the "action" query
         *     on the alarms named in the JSON string array in the body.  The action
         *     can be performed on alarms across multiple namespaces, including non-namespaced
         *     alarms.  It is possible for some actions on alarms to have been successful and
         *     other actions to have failed if an error is returned. */
        put: operations["alarmPutAlarms"];
        post?: never;
        /** @description Deletes all the alarms named in the JSON string array in the body.  Deletes
         *     can be performed on alarms across multiple namespaces, including non-namespaced
         *     alarms.  It is possible for some deletes on alarms to have been successful and
         *     other deletes to have failed if an error is returned. */
        delete: operations["alarmDeleteAlarms"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        NamespaceGetResponse: {
            /** @description If true, the requestor is considered to have permission to access all namespaces */
            allNamesapces?: boolean;
            /** @description The list of namespaces */
            namespaces?: components["schemas"]["NamespaceData"][];
        };
// ...
        TransactionDetails: {
            /** @description List of changed CRs as part of the transaction */
            changedCrs?: components["schemas"]["TransactionNsCrGvkNames"][];
            /** @description True if the transaction was not committed */
            dryRun?: boolean;
            /** @description List of general errors while running the transaction */
            generalErrors?: string[];
            /** @description List of input CRs from the transaction */
            inputCrs?: components["schemas"]["TransactionInputResource"][];
            /** @description List of intents which ran as part of the transaction */
            intentsRun?: components["schemas"]["TransactionIntentResult"][];
            /** @description List of nodes with configuration changes from the transaction */
            nodesWithConfigChanges?: components["schemas"]["TransactionNodeResult"][];
            /** @description The state of the transaction */
            state?: string;
            /** @description True if the CR was successfully applied */
            success?: boolean;
        };
// ...
        TransactionSummaryResults: {
            /** @description array of summary-results for transactions */
            results?: components["schemas"]["TransactionSummaryResult"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}